Updated fetchArticles model to return comment_count as number. Updted tests to reflect this

+++ models/articles_models.js
@@ -11,7 +11,7 @@ async function fetchArticles(query, queryNames) {
 
     let sqlStr = `SELECT articles.author, articles.title, articles.article_id,
     articles.topic, articles.created_at, articles.votes, articles.article_img_url, 
-    COUNT(comments.comment_id) AS comment_count
+    COUNT(comments.comment_id)::INT AS comment_count
     FROM articles
     LEFT JOIN comments
     ON articles.article_id = comments.article_id `
